changed time to epoch
