--- log-data/log-data
@@ -1,467 +1,0 @@
-=~=~=~=~=~=~=~=~=~=~=~= PuTTY log 2023.03.07 13:44:21 =~=~=~=~=~=~=~=~=~=~=~=
-pressure: 84810
-altitude: 1475.71
-x: -0.08
-y: 0.23
-z: 10.36
-pressure: 84808
-altitude: 1476.00
-x: -0.07
-y: 0.23
-z: 10.37
-pressure: 84807
-altitude: 1476.09
-x: -0.07
-y: 0.23
-z: 10.38
-pressure: 84807
-altitude: 1475.71
-x: -0.08
-y: 0.24
-z: 10.36
-pressure: 84808
-altitude: 1475.71
-x: -0.08
-y: 0.24
-z: 10.37
-pressure: 84810
-altitude: 1475.81
-x: -0.07
-y: 0.23
-z: 10.37
-pressure: 84814
-altitude: 1475.81
-x: -0.06
-y: 0.24
-z: 10.38
-pressure: 84806
-altitude: 1475.52
-x: -0.07
-y: 0.23
-z: 10.36
-pressure: 84808
-altitude: 1475.71
-x: -0.08
-y: 0.23
-z: 10.37
-pressure: 84813
-altitude: 1475.52
-x: -0.06
-y: 0.23
-z: 10.36
-pressure: 84811
-altitude: 1476.00
-x: -0.07
-y: 0.23
-z: 10.36
-pressure: 84812
-altitude: 1475.61
-x: -0.07
-y: 0.23
-z: 10.37
-pressure: 84808
-altitude: 1475.90
-x: -0.07
-y: 0.23
-z: 10.36
-pressure: 84813
-altitude: 1475.61
-x: -0.07
-y: 0.23
-z: 10.36
-pressure: 84803
-altitude: 1475.61
-x: -0.07
-y: 0.23
-z: 10.36
-pressure: 84804
-altitude: 1476.19
-x: -0.06
-y: 0.23
-z: 10.37
-pressure: 84809
-altitude: 1475.61
-x: -0.07
-y: 0.23
-z: 10.37
-pressure: 84809
-altitude: 1475.71
-x: -0.07
-y: 0.23
-z: 10.36
-pressure: 84808
-altitude: 1475.61
-x: -0.07
-y: 0.23
-z: 10.38
-pressure: 84816
-altitude: 1475.61
-x: -0.06
-y: 0.23
-z: 10.37
-pressure: 84811
-altitude: 1475.71
-x: -0.06
-y: 0.23
-z: 10.35
-pressure: 84807
-altitude: 1475.61
-x: -0.07
-y: 0.23
-z: 10.36
-pressure: 84806
-altitude: 1475.81
-x: -0.06
-y: 0.23
-z: 10.36
-pressure: 84807
-altitude: 1476.09
-x: -0.06
-y: 0.23
-z: 10.36
-pressure: 84811
-altitude: 1475.52
-x: -0.06
-y: 0.24
-z: 10.36
-pressure: 84805
-altitude: 1476.19
-x: -0.06
-y: 0.23
-z: 10.37
-pressure: 84806
-altitude: 1475.81
-x: -0.08
-y: 0.23
-z: 10.36
-pressure: 84807
-altitude: 1476.19
-x: -0.06
-y: 0.23
-z: 10.37
-pressure: 84809
-altitude: 1475.42
-x: -0.07
-y: 0.22
-z: 10.36
-pressure: 84811
-altitude: 1475.81
-x: -0.07
-y: 0.23
-z: 10.39
-pressure: 84808
-altitude: 1475.61
-x: -0.07
-y: 0.24
-z: 10.37
-pressure: 84811
-altitude: 1475.81
-x: -0.07
-y: 0.23
-z: 10.37
-pressure: 84807
-altitude: 1476.09
-x: -0.08
-y: 0.24
-z: 10.37
-pressure: 84807
-altitude: 1476.00
-x: -0.07
-y: 0.23
-z: 10.38
-pressure: 84810
-altitude: 1476.00
-x: -0.07
-y: 0.23
-z: 10.37
-pressure: 84809
-altitude: 1475.81
-x: -0.05
-y: 0.24
-z: 10.37
-pressure: 84808
-altitude: 1476.00
-x: -0.20
-y: 0.28
-z: 10.38
-pressure: 84810
-altitude: 1475.42
-x: -0.17
-y: 0.27
-z: 10.37
-pressure: 84809
-altitude: 1476.00
-x: -0.14
-y: 0.25
-z: 10.35
-pressure: 84810
-altitude: 1475.71
-x: -1.12
-y: -0.09
-z: 9.65
-pressure: 84809
-altitude: 1475.81
-x: -2.01
-y: 0.77
-z: 11.38
-pressure: 84811
-altitude: 1475.90
-x: 0.42
-y: -2.29
-z: 14.81
-pressure: 84809
-altitude: 1476.00
-x: 5.97
-y: 1.61
-z: -1.77
-pressure: 84810
-altitude: 1475.90
-x: 6.57
-y: 2.21
-z: -7.43
-pressure: 84807
-altitude: 1476.19
-x: 4.66
-y: 0.76
-z: -8.45
-pressure: 84805
-altitude: 1475.71
-x: 4.73
-y: 1.56
-z: -8.62
-pressure: 84804
-altitude: 1475.71
-x: 4.23
-y: 1.71
-z: -8.90
-pressure: 84802
-altitude: 1475.81
-x: 3.36
-y: 2.09
-z: -8.77
-pressure: 84813
-altitude: 1475.81
-x: 3.12
-y: 2.00
-z: -9.64
-pressure: 84808
-altitude: 1475.71
-x: 3.41
-y: 2.33
-z: -8.75
-pressure: 84810
-altitude: 1475.90
-x: 3.61
-y: 2.33
-z: -9.05
-pressure: 84809
-altitude: 1475.61
-x: 3.95
-y: 2.41
-z: -8.32
-pressure: 84808
-altitude: 1475.52
-x: 5.33
-y: 1.65
-z: -8.34
-pressure: 84809
-altitude: 1475.90
-x: 5.66
-y: 1.73
-z: -7.59
-pressure: 84813
-altitude: 1475.90
-x: 5.84
-y: 1.56
-z: -8.03
-pressure: 84811
-altitude: 1476.00
-x: 5.54
-y: 1.71
-z: -8.10
-pressure: 84808
-altitude: 1475.71
-x: 5.44
-y: 1.83
-z: -8.13
-pressure: 84812
-altitude: 1475.71
-x: 5.89
-y: 1.87
-z: -7.87
-pressure: 84804
-altitude: 1475.71
-x: 7.66
-y: 2.03
-z: -8.58
-pressure: 84808
-altitude: 1475.90
-x: 7.88
-y: 2.95
-z: 1.44
-pressure: 84808
-altitude: 1475.90
-x: 6.72
-y: 4.59
-z: 5.80
-pressure: 84809
-altitude: 1475.81
-x: 3.67
-y: 3.37
-z: 10.31
-pressure: 84810
-altitude: 1475.81
-x: 1.69
-y: 5.76
-z: 7.55
-pressure: 84808
-altitude: 1476.19
-x: -8.16
-y: 3.46
-z: 8.68
-pressure: 84814
-altitude: 1475.32
-x: -9.45
-y: 3.26
-z: 2.03
-pressure: 84813
-altitude: 1475.52
-x: -7.85
-y: 1.57
-z: 1.98
-pressure: 84809
-altitude: 1475.32
-x: -9.33
-y: 1.70
-z: 2.05
-pressure: 84817
-altitude: 1475.32
-x: -9.32
-y: 1.76
-z: 1.79
-pressure: 84816
-altitude: 1476.00
-x: -9.39
-y: 1.72
-z: 1.89
-pressure: 84814
-altitude: 1475.61
-x: -9.36
-y: 1.79
-z: 1.70
-pressure: 84814
-altitude: 1475.32
-x: -9.40
-y: 1.85
-z: 1.63
-pressure: 84811
-altitude: 1475.23
-x: -9.34
-y: 1.86
-z: 1.60
-pressure: 84812
-altitude: 1475.52
-x: -9.38
-y: 1.81
-z: 1.68
-pressure: 84817
-altitude: 1475.81
-x: -9.36
-y: 1.78
-z: 1.72
-pressure: 84809
-altitude: 1475.42
-x: -9.40
-y: 1.80
-z: 1.66
-pressure: 84812
-altitude: 1475.81
-x: -9.26
-y: 1.77
-z: 1.70
-pressure: 84816
-altitude: 1475.42
-x: -9.40
-y: 1.85
-z: 1.61
-pressure: 84808
-altitude: 1475.13
-x: -10.30
-y: 0.25
-z: 1.19
-pressure: 84815
-altitude: 1475.52
-x: -6.49
-y: -3.54
-z: 6.23
-pressure: 84812
-altitude: 1475.81
-x: -2.99
-y: -5.95
-z: 7.57
-pressure: 84812
-altitude: 1475.90
-x: -1.85
-y: -5.86
-z: 7.87
-pressure: 84814
-altitude: 1475.81
-x: -1.12
-y: -5.58
-z: 8.92
-pressure: 84810
-altitude: 1475.81
-x: -0.90
-y: -1.34
-z: 9.22
-pressure: 84807
-altitude: 1476.09
-x: -2.34
-y: 0.76
-z: 11.73
-pressure: 84808
-altitude: 1475.90
-x: -2.07
-y: 2.01
-z: 9.87
-pressure: 84807
-altitude: 1475.90
-x: -2.22
-y: 2.42
-z: 9.97
-pressure: 84809
-altitude: 1476.00
-x: -2.25
-y: 2.54
-z: 9.90
-pressure: 84809
-altitude: 1475.52
-x: -2.48
-y: 2.50
-z: 9.80
-pressure: 84814
-altitude: 1475.81
-x: -2.43
-y: 2.73
-z: 9.64
-pressure: 84804
-altitude: 1476.09
-x: -2.63
-y: 3.36
-z: 9.36
-pressure: 84810
-altitude: 1475.90
-x: -2.35
-y: 3.49
-z: 9.20
-pressure: 84808
-altitude: 1476.09
-x: -2.56
-y: 3.81
-z: 9.15
-pressure: 84809
-altitude: 1475.81
-x: -2.31
-y: 3.78
-z: 9.31
-p
